Unified pagefinder integration to handle both prod and dev
