+++ astro.config.mjs
@@ -1,8 +1,8 @@
 import { defineConfig } from "astro/config";
-import { pagefindDev } from "./src/integrations/pagefind-dev.ts";
+import pagefind from "./src/integrations/pagefind.ts";
 
 export default defineConfig({
-  integrations: [pagefindDev()],
+  integrations: [pagefind()],
   redirects: {
     "/avdelinger": "/avdelinger/trondheim",
   },
